fix: ignore extra data in discv4 messages (#110)
**Motivation**

According to the spec:
> Implementations should ignore any additional elements in the
`packet-data` list as well as any extra data after the list.

We're currently failing if any of that happens.

**Description**

This PR introduces `Decoder::finish_unchecked`, which ignores any
trailing bytes instead of failing, and uses that to allow for additional
elements inside the packet-data. Also, it replaces some usage of
`RLPDecode::decode` with `RLPDecode::decode_unfinished` to allow for
extra data after the message.

As an extra, it refactors out the big `match` inside of both `Message`
encode and decode functions, since it's expected those will change often
in the next PRs.

Related to #17