feat: add RLP validations on EF tests (#157)
**Motivation**

**Description**
Validate RLP data before executing the EF tests.
Introduces the following changes
* `Block` struct along with the RLP encoding and decoding traits
* Conversions from `ef_tests` to `core` types (`Block`, `Transaction`,
`Withdrawal`)
* Changes how non-legacy transactions (EIP1559, EIP2930, EIP4844) are
encoded and decoded. These were being encoded as a list of fields but
with this PR they are encoded as a single RLP item whose payload is
`tx_type || rlp_encoded_tx`.
* Set `BlockHeader` pre-Cancun fields
(`parent_beacon_block_root`,`blob_gas_used`, `excess_blob_gas`) and
`withdrawals_root` (some of the ef_tests do not include this field) as
optional.
* Validates the genesis block by decoding it and comparing its header to
the deserialized one.
* Validates each block by decoding it, comparing it to the deserialized
one, and encoding it and comparing the encoding to the `rlp` field.

**Some errors I found**
* When executing the `ef_tests` some will fail with `"lack of funds for
max fee"`. As far as I know this is because before executing the
transaction, the `sender` field must be recovered (from the message and
its signature) and apparently it is not being recovered properly (it
does not match with the `sender` field included in the deserialized
`ef_tests::Transaction`).
 
Closes #155