feat: eip4788 update beacon block root cotract before processing block transactions (#211)
**Motivation**

Implements
[eip4788](https://eips.ethereum.org/EIPS/eip-4788#specification), the
eip4788 beacon root contract must be called before every transaction to
update it with the latest consensus state.

**Description**

1. Implements the following function to perform the contract system call
before block execution:

```rust
pub fn beacon_root_contract_call(
    state: &mut EvmState,
    header: &BlockHeader,
    spec_id: SpecId,
) -> Result<ExecutionResult, EvmError>
```
This contract call must not increase the system-address's nonce, to
achieve this, the transaction is executed with the `transact` function,
then, the after-execution state it is modified to remove the changes to
the system address account and to the coinbase and finally the changes
are commited to the database.

2. Fixes `increase_balance` in the store engine, when this method is
called with an account that does not exist in the database, the account
must be added to the database instead of ignoring the call.

3. Implements `execute_block` function to abstract the functionality
that was being developed in `execute_test`.

4. Implements `test_execute_block` function which does the same as
}execute_test` but using the `execute_block` function, does not test for
exceptions yet.

This gets the following test-suites to pass when checking the test's
post-state and the achieved post-state:
- vectors/cancun/eip4788_beacon_root
- vectors/cancun/eip1153_tstore
- vectors/cancun/eip5656_mcopy

Closes https://github.com/lambdaclass/ethereum_rust/issues/27