feat: pass SpecId to tests and to execute_block (#229)
**Motivation**

Many tests were not passing because the execution spec passed to the EVM
did not match with the specified fork in the test.

For example, tests with self-destructing contracts and Shanghai spec
where not passing with the hard-coded Cancun spec, because the EVM's
behaviour for this opcode is different in these two forks.

**Description**

Parses the "network" value from the tests, and passes the fork spec to
the EVM.
Tests which specify a network different than Cancun or Shanghai are
ignored for now.

With this changes, all tests from the Shanghai test-suite are passing
and all tests from the Cancun test-suite except the ones from
eip4844_blobs.

fixes: https://github.com/lambdaclass/ethereum_rust/issues/99
closes: https://github.com/lambdaclass/ethereum_rust/issues/228