fix: set chain_id in evm config before running transactions (#234)
**Motivation**

Some transactions fail when running with kurtosis with the evm error
"invalid chain id". This is because the evm uses 1 as default chain id
when it is not set and then checks that the transaction's chain_id
matches the one in its config when validating the transaction.

**Description**

* Fetches chain_id from the state and sets it in the evm config before
running transactions
* Misc: Fixes warning on Dockerfile
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes None, but is needed for #51