fix: use canonical transaction encoding for transactions in payloads and root calculations (#240)
**Motivation**

In a previous PR, transaction encoding and decoding for non legacy
transactions was changed from `tx_type || rlp(tx_fieldss)` to
`rlp(bytes(tx_type || rlp(tx_fields)))`. While this is the correct
encoding used when encoding and decoding a block with transactions, it
removed the implementation of the "canonical" encoding that is used when
decoding an execution payload and computing transaction roots. This PR
aims to fix this issue by implementing the previous canonical encoding
separately from the rlp encoding and updating the doc to avoid future
confusions and problems

<!-- Why does this pull request exist? What are its goals? -->

**Description**

* Implement `encode_canonical` & `decode_canonical` for Transaction and
use it to decode transactions in execution payloads and encode
transactions when computing the transactions root.
* Update documentation for rlp implementation of transactions to clearly
state that it is encoding the transaction into a bytes object.

* Other:
* Revert rlp encoding of inner transactions to only encode the
transaction fields and move type and bytes encoding to the
implementation of encoding for the transaction enum. Reasoning: This
allows us to use the rlp encoding for the inner transactions for both
the canonical and bytes encoding of the transaction enum, and also fixes
rlpdecode and rlpencode not being consistent with one another)

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes None, but is needed for #51 to work

With this fix we can now successfully validate and execute all payloads
with kurtosis local network 🚀



https://github.com/user-attachments/assets/9d977321-00d8-467c-ab38-a97b62a8a2cc