feat: eip-4844 block validation (#239)
**Motivation**

Validate blocks before execution to correctly execute eip4844
transactions and blocks.

**Description**

Introduces the `validate_block` function which performs pre-execution
checks for validity of the block in reference to the parent_header and
the chain config and current fork.

`validate_block` returns a boolean value which is true when the block
can be safely executed.
This does not mean the block execution won't fail due to insufficient
gas for a transaction.

Using this function before execution gets all `eip-4844` tests to pass
by filtering invalid blocks which before where being executed without a
failure, leading to inconsistent chain state.

Tests where also failing due to a small issue in the `block_env`
function where the blob fields of the header where not being set, this
is now fixed.

Closes https://github.com/lambdaclass/ethereum_rust/issues/227

---------

Co-authored-by: ElFantasma <estebandh@gmail.com>