fix: `eth_getBalance`, `eth_getCode` & `eth_getStorageAt`: return default value if account is not found (#259)
**Motivation**

Rpc endpoint `eth_getBalance` should return 0 if the account is not
present in the storage.
Rpc endpoint `eth_getCode` should return empty bytes if the account is
not present in the storage.
Rpc endpoint `eth_getStorageAt` should return 0 if the account or
storage slot is not present in the storage.
<!-- Why does this pull request exist? What are its goals? -->

**Description**

* `eth_getBalance`: return zero if the account is not part of the state
* `eth_getCode`: return empty bytes if the account is not part of the
state
* `eth_getStorageAt`: return zero if the account or storage slot is not
part of the state
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes None, but makes hive test
`eth_getBalance/get-balance-unknown-account` &
`eth_getCode/get-code-unknown-account` pass.
(`eth_getStorageAt/get-storage-unknown-account` would also pass if it
weren't for formatting differences)