fix: update latest block number when adding a block (#255)
**Motivation**

When we add a block to this store this block will always become the
latest block.
Previously we handled this by updating the latest block number after we
added a new block but this was not consistent as we had to remember to
update the last block number after each call to add block, leading to
bugs.
This PR aims to solve this issue by including the update of the latest
block number in the store's add block functionality.

**Description**

* Call `update_latest_block_number` in `Store::add_block`
* Remove calls to `update_latest_block_number` elsewhere
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes Closes #266