refactor: extract functionality into higher level chain crate  (#257)
**Motivation**

Create a new crate chain that manages the blockchain from a high level
perspective.

**Description**
Removes the empty consensus crate.

Introduces the `chain` crate with the `add_block` function, which takes
a storage and a block, performs the preliminary validations, executes
the transactions in the block and updates the storage.

This is a proposal for implementing the crate specified in the issue
https://github.com/lambdaclass/ethereum_rust/issues/251, since this is
more of a design issue, feedback and suggestions are very much
appreciated.