feat: jwt authentication for RPC  (#302)
**Motivation**

Json RPC calls to the engine execution API should be authenticated by
bearing a JWT token as specified
https://github.com/ethereum/execution-apis/blob/main/src/engine/authentication.md

This is to validate that authrpc calls are only issued by the consensus
layer and prevents attack which would come from accidentally exposing
the execution client to the internet.

**Description**

Introduces authentication.rs module which uses the jsonwebtoken crate to
decode and validate tokens issued by the consensus layer.

The tokens contain a "iat" claim which stands for "issued at timestamp",
according to the spec, this unix timestamp must be at most 60 seconds
from the time of validation.

For this PR to pass the CI tests, this one needs to be merged first on
lambdaclass/hive to enable the testing of the authentication by Hive:
https://github.com/lambdaclass/hive/pull/2/files

To enable authentication on our Kurtosis localnet this needs to be
merged:
https://github.com/lambdaclass/ethereum-package/pull/2/files


Closes https://github.com/lambdaclass/ethereum_rust/issues/13