feat: cache built dependencies on Dockerfile using cargo chef (#306)
**Motivation**

Building the container image for the client now takes up to a minute and
a half each time.

This makes testing the client in Kurtosis or with Hive really tedious.

Most of the build time is wasted building the dependencies, [cargo
chef](https://hub.docker.com/r/lukemathwalker/cargo-chef) is a tool for
speeding up cargo docker files by building the dependencies and the
final binary in two different steps, allowing for the dependencies to be
cached by the docker engine.

**Description**
In a Dockerfile, everything before a COPY statement is cached if there
where no changes to the files being copied. This can be leveraged to
reduce building times.

Modifies the Dockerfile to use cargo chef in a multi-stage build,
caching the dependencies and significantly speeding up the image's build
time.

_Build time before this PR: 92 seconds_

![image](https://github.com/user-attachments/assets/ef60204c-7253-4447-9361-79d3be714e64)

_Build time running for the first time after this PR: 102 seconds_

![image](https://github.com/user-attachments/assets/305bc81f-dc46-4807-9861-9c21e7f4c55b)

_Build time for consequent builds after chaching dependencies: 15
seconds_

![image](https://github.com/user-attachments/assets/79a40d4f-41e0-44c0-bfb2-f9bbc1d91cb5)


Closes none