fix: get_block_by_number endpoint (#310)
**Motivation**

The `get-finalized` and `get-safe` endpoints were returning null values
because we weren't handling the forkchoice updates from the engine.

**Description**

This PR adds a minimal implementation to store forkchoice updates and
pass the tests. Additionally, it takes the opportunity to implement a
refactor.

- Moves `core/types/engine/payload.rs` to `rpc/types/payload.rs`.
- Creates `rpc/engine/payload.rs` and `rpc/engine/fork_choice.rs` to
modularize `rpc/engine/mod.rs`.
- Adds a minimal implementation of `forkchoice_updated_v3()`.
- Fixes`eth_getBlockByNumber`.

**Observations**

This PR doesn't implement the `engine_forkchoiceUpdatedV3` endpoint, It
only stores the minimal data needed to pass `eth_getBlockByNumber`.

Closes #270