feat: admin_NodeInfo endpoint (#312)
**Motivation**

The RPC method `admin_nodeInfo` of the admin namespace is a non standard
endpoint that might be necessary for configuring the hive p2p tests, it
provides information about the node's network and protocol
configuration.

**Description**

Introduces the RPC `admin_nodeInfo` endpoint which is only accessible
through the HTTP API. It returns the enode configuration and, for the
eth protocol, the chain config that was set on genesis.
To achieve this, a `ChainConfig` struct and a `Node` struct representing
the eth and p2p configurations respectively must be passed to the http
request handler so that it can provide the correct response.
This data is passed to the handler by adding it to the `RpcApiContext`
struct .

Introduces a small refactor to the net crate, extracting the types
`Endpoint` `Node` and `NodeRecord` to a public `types` module.

Contains a small rewrite to the RPC routing of requests, separating them
by namespace.

Example output:
```bash
➜  ethereum_rust git:(admin-nodeinfo) curl -s -X POST -H "Content-Type: application/json" --data '{"jsonrpc":"2.0", "method":"admin_nodeInfo", "params":[], "id":1}' "localhost:59196"


{
   "id":1,
   "jsonrpc":"2.0",
   "result":{
      "enode":"enode://f76e29b3291f595ac9e0465250e6e4e1cc57f8d2b02503d52b5e5175739878ae6d6222d38b777b0f89888ad65806d0c97a8643b54b67d161201aa43a88765d34@0.0.0.0:30303?discport=30303",
      "id":"f76e29b3291f595ac9e0465250e6e4e1cc57f8d2b02503d52b5e5175739878ae6d6222d38b777b0f89888ad65806d0c97a8643b54b67d161201aa43a88765d34",
      "ip":"0.0.0.0",
      "name":"ethereum_rust/0.1.0/rust1.80",
      "ports":{
         "discovery":30303,
         "listener":30303
      },
      "protocols":{
         "eth":{
            "chainId":"0x301824",
            "homesteadBlock":0,
            "daoForkBlock":null,
            "daoForkSupport":false,
            "eip150Block":0,
            "eip155Block":0,
            "eip158Block":0,
            "byzantiumBlock":0,
            "constantinopleBlock":0,
            "petersburgBlock":0,
            "istanbulBlock":0,
            "muirGlacierBlock":null,
            "berlinBlock":0,
            "londonBlock":0,
            "arrowGlacierBlock":null,
            "grayGlacierBlock":null,
            "mergeNetsplitBlock":0,
            "shanghaiTime":0,
            "cancunTime":0,
            "pragueTime":1725062017,
            "verkleTime":null,
            "terminalTotalDifficulty":"0x0",
            "terminalTotalDifficultyPassed":true
         }
      }
   }
}%
```


<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes https://github.com/lambdaclass/ethereum_rust/issues/248

---------

Co-authored-by: Avila Gastón <72628438+avilagaston9@users.noreply.github.com>