refactor: adds TrieDB trait + Libmdbx implementation (#389)
**Motivation**

As explained in #359 this is a step towards decoupling the Trie logic
from the storage strategy. It's important if we want to merge this code
back to the original trie codebase, which is an independent library.

**Changes**

- Added TrieDB trait.
- Added Libmdbx implementation for the TrieDB trait.
- TrieState and Trie now receive a TrieDB argument instead of a path.
- Updated all structs and tests to use the generic.

**Not included in this PR**

- We could take TrieState out too, by making the Cache a TrieDB
implementation too, that can have a db as well.
- Libmdbx has a Nodes table hardcoded for now. Ideally, it should
receive a parameter to have a table for each trie, to separate, for
instance, the transaction trie and the transaction receipt trie.