refactor: allow different tables Libdbmx TrieDB implementation using generics (#397)
**Motivation**

This PR allows us to have multiple tries, all using the same database,
but different tables, so they can't collide with each other even if they
were to share key space.

**Changes**

- Makes the `Libmdbx` implementation for the `TrieDB` trait generic. The
type is a Table struct which specifies which table will be used in the
trie.
- The `Libmdbx` struct now has a borrowed db instead of a moved one.
This allows us to use a single db struct in several tries, which didn't
work in the previous model (`resource temporarily unavailable`).
- Adds `test_utils` for creating a db. Most of the PR code is just
changing tests, but the Trie code itself is 100% unchanged.

Closes #359