feat: use project dir as default DB folder
Adds `directories-rs` as dependency. This is used to find the approriate
directory for the project depending on the OS.
Uses the local data directory for the project as the default place to
store a libmdbx storage if none is passed. This is handled manually
during initialization rather than by clap due to lifetime issues, as
`Command` tries to store a reference rather than an owned object.