feat: add basic `mempool` and `eth_sendRawTransaction` RPC endpoint (#408)
**Motivation**

We need a mempool where we can store all transactions received in a node
which are waiting to be included in a new block.
Those transactions may be received by three possible entry points, so we
need to implement at least one of them.

**Description**

This PR implements `eth_sendRawTransaction` RPC endpoint, as well as a
very basic first version of a transaction mempool which stores those
transactions on the storage. Those transactions will have to be
validated before they can be added to the pool, which is something that
is not addressed in this PR.

Closes:

#348 
#175 
#336

---------

Co-authored-by: Julian Ventura <julian.ventura@lambdaclass.com>