feat: remove `Mutex` from `Store` (#441)
**Motivation**
Currently, our Store contains an `Arc<Mutex<dyn StoreEngine>>`, but the
mutex is not really needed for the libmbdx backend, it is only needed
for the InMemory one. This means we should be able to move the mutex
into the in memory engine and leave the libmdbx mutex-free.
This PR proposes moving this mutex into the InMemory engine and changing
the structure of `Store` to:
```
pub struct Store {
    engine:  Arc<dyn StoreEngine>
}
```


<!-- Why does this pull request exist? What are its goals? -->

**Description**

* Change `Store.engine` type from `Arc<Mutex<dyn StoreEngine>>` to ` E:
StoreEngine`,
* Wrap in memory store engine in `Arc<Mutex<>>`
* Rename implementations of `StoreEngine` to `LibmdbxStoreEngine` &
`InMemoryStoreEngine` and expose them publicly
* Add method `new_temp` for `Store` to be used in tests instead of
creating and removing directories "by hand".

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number