feat: add transaction validations to `mempool` (#440)
**Motivation**

When a new transaction is going to be added to the mempool, first we
need to run some very basic validations so we can easily discard those
that would fail later on.

**Description**

This PR adds some transaction validations to the mempool, together with
unit tests to ensure their correct behaviour.

---------

Co-authored-by: Julian Ventura <julian.ventura@lambdaclass.com>