feat: improve revalidation with liveness checks (#431)
**Motivation**

Implements liveness in peers for revalidations.

**Description**

This pr introduces an improvement over the current peer revalidations.
This new form pings peers more often (every 30 secs) and keeps track of
the ping answers, which provides generally good criteria to prefer some
nodes over others when interacting in the `p2p` and choosing the initial
seeders (or bootnodes see #398).

Here is how the new revalidation flow works:
1. Every `30` seconds (by default) we ping the three least recently
pinged peers: this may be fine now to keep simplicity, but we might
prefer to choose three random peers instead to avoid the search which
might become expensive as our buckets start to fill with more peers.
2.  In the next iteration we check if they have answered
     -  if they have: we increment the liveness field by one
     - otherwise: we decrement it by the livenes by `/ 3`.
3. If the liveness field is `0`, we delete it and insert a new one from
the replacements table.

Closes #417