docs: update milestones (#446)
**Description**
Changed the ordering of milestones.

I've lowered the P2P milestones since it's not really needed if we want
to use the client for rollups. Apart from this, block building with
enable us to run the `engine-api` Hive simulation, which is key to tests
the client, specially when it comes to reorgs.