feat: support reorgs when parent is canonical (#445)
**Motivation**

Right now, no support for reorgs is available. We assume that the new
blocks and heads are direct successors of the current head. This PR adds
support for non-canonical blocks to be added, assuming their parents are
part of the canonical chain.

**Changes**

- Engine API `NewPayload`:
- Instead of validating that the new block is a child of the head, it
validates that the parent is any canonical block.
  - Remove fork choice update.
- `ForkChoiceUpdated`: 
- Validates that the head's parent, the safe and finalized blocks are
canonical. Returns an invalid response if not.
- Applies the fork choice update changing the canonical block at the
head's height.

Closes #350