feat: (levm) Implement CALL and RETURN opcodes (#568)
**Motivation**
Implemented de `CALL` and `RETURN` opcodes, which introduce the logic of
moving between `callframes`.

**Description**
- The `callframes` vec now acts as a stack, by popping and pushing new
callframes. The current callframe is popped and its parents are the ones
stored in the vec, the last callframe in the vec is the current
callframes' parent.
- A new `Account` struct is created, each account stores (for now) its
balance and bytecode. The `VM` now holds a hashmap to store accounts.

Closes #529 
Closes #531