feat(L2): add basic proof data provider (#585)
**Motivation**

We need to have a ProofDataProvider that serves as an interface to
communicate with the prover.
<!-- Why does this pull request exist? What are its goals? -->

**Description**

The ProofDataProvider listen 2 types of messages: ProofData Requests and
ProofData Submits. The first ones are asked by the prover to start a new
proof job. The second one is used by the prover to inform that the proof
generation is ready.
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

---------

Co-authored-by: ilitteri <ilitteri@fi.uba.ar>