feat(levm): transaction environment (#582)
**Motivation**

To implement the environment in which the transactions run and the call
frames are created.

**Description**

Implementation of transaction structures and transaction environment to
store the executing transaction data.

**Changelist**
- Create a `Message` struct that contains the necessary information for
the execution of a `CallFrame`.
- Modify `CallFrame` interface to be created with a `Message`.
- Rename `Db` struct to `Storage`.
- Add `WorldState` struct to hold the mapping between addresses and
accounts.
- Complete `Account` struct to have `nonce` and `Storage`. Modify tests
to access the storage through the `WorldState`.
- Create `Environment` struct to hold the info that should be globally
accessible to all executing call frames: origin address, block
environment, etc.
- Create `Substate` unit like struct as a placeholder for the _accrued
substate_ data: touched accounts, accounts to delete, etc.
- Modify `VM` struct to hold the `Environment`, the `Substate` and the
`WorldState`.

---------

Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>