refactor: improve block header validation error handling (#699)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

Nowadays important block validation errors are omitted. If a block is
added to the blockchain and fails at header validation you have no clue
which of all validations failed, only that the error was a
`ChainError::InvalidBlock(InvalidBlockError::InvalidHeader)`.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

This PR proposes an improvement in the handling of block validation
errors.