fix: implementation of `TxKind::deserialize` (#706)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

The actual implementation of `TxKind` (the field `to` of a transaction)
expects a `String`. But this field's value can be `null` which does not
match with the `String` type and breaks the deserialization.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

This PR fixes the `TxKind` deserialization by expecting an
`Option<String>` instead of a `String` and the deserialized value will
be `TxKind::Create` if there's an empty `String` or a `null` and `Call`
if the incoming `String` can be parsed as an `Address`.

---------

Co-authored-by: fmoletta <99273364+fmoletta@users.noreply.github.com>