feat: (levm) Update error handling  (#705)
**Summary**

Delete the unwraps/expects from the `VM` execution of bytecode.
We previously handled the moving between calls with `callframes` by just
changing the current callframe being executed.
Now we have added the `ExecutionResult` to be returned by `execute`,
this adds the possibility to obtain the specifics of what happened
during the execution of a certain callframe. This affects mainly the way
`CALL` and `CREATE` functions work. With this PR we start calling into
new callframes in a recursive way to better handle the result of each
execution.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
- Adds the `Stack` struct for handling the pop/push errors
- Returns `Halts` as error
- Changed the `CALL` and `CREATE` opcode to do a recursive call to
execute
- Execute returns a `ExecutionResult` with information
- Added the types `Success`, `Revert` and `Halt` for `ExecutionResult`. 
- Added support for the revert function in `generic_call`.
- Added a revert if an opcode passes the gas limit.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #539 
Closes #556