feat(levm): add Unit Tests for Various Opcodes (#696)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
The objective is to add at least one unit test for every opcode, so that
if we want to make changes in that opcode we can run it's own test and
see if it breaks.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
Implements tests for every opcode of the list down below.
When making tests for opcode `SMod` I realized it wasn't working
properly for negative numbers. As it wasn't a huge change I fixed it in
this PR.

### Opcodes Tests
•	Mul 
•	Sub
•	Div
•	Sdiv
•	Mod
•	`SMod`
•	Addmod
•	Mulmod
•	Exp
•	SignExtend
•	Lt
•	Gt
•	Slt
•	Sgt
•	Eq
•	IsZero

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #682 and #704 


### Additional Note
When merging main branch into this I realized that some changes that I
had made previously were reverted. I guess there was a misunderstanding
when solving conflicts in EF Tests Branch.
I applied the changes once again, it is using `ExecutionResult` instead
of `Result<ExecutionResult,VMError>` in `execute()` method of vm.

I also removed an unused parameter in `execute()`. If necessary we can
add it later.